[PAGOPA-2745] refactoring packages
