+++ src/main/java/it/gov/pagopa/fdr/conversion/model/ErrorEnum.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.conversion.util;
+package it.gov.pagopa.fdr.conversion.model;
 
 public enum ErrorEnum {
     GENERIC_ERROR,
